# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shakti Fast Food API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuItem {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  halfPrice?: number | null;
  category: string;
  /** @nullable */
  image?: string | null;
  available: boolean;
  isVeg: boolean;
  isFeatured?: boolean;
  orderCount?: number;
}

export interface MenuItemInput {
  name: string;
  description?: string;
  price: number;
  /** @nullable */
  halfPrice?: number | null;
  category: string;
  /** @nullable */
  image?: string | null;
  available?: boolean;
  isVeg: boolean;
  isFeatured?: boolean;
}

export interface MenuItemUpdate {
  name?: string;
  description?: string;
  price?: number;
  /** @nullable */
  halfPrice?: number | null;
  category?: string;
  /** @nullable */
  image?: string | null;
  available?: boolean;
  isVeg?: boolean;
  isFeatured?: boolean;
}

export interface AvailabilityUpdate {
  available: boolean;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  icon?: string | null;
  itemCount?: number;
}

export interface CategoryInput {
  name: string;
  slug: string;
  /** @nullable */
  icon?: string | null;
}

export interface OrderItem {
  menuItemId: number;
  name: string;
  price: number;
  quantity: number;
  /** @nullable */
  size?: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  received: 'received',
  preparing: 'preparing',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type OrderOrderType = typeof OrderOrderType[keyof typeof OrderOrderType];


export const OrderOrderType = {
  'dine-in': 'dine-in',
  takeaway: 'takeaway',
} as const;

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  orderType: OrderOrderType;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tableNumber?: string | null;
  createdAt: string;
}

export type OrderInputOrderType = typeof OrderInputOrderType[keyof typeof OrderInputOrderType];


export const OrderInputOrderType = {
  'dine-in': 'dine-in',
  takeaway: 'takeaway',
} as const;

export interface OrderInput {
  customerName: string;
  phone: string;
  items: OrderItem[];
  orderType: OrderInputOrderType;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tableNumber?: string | null;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  received: 'received',
  preparing: 'preparing',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AuthToken {
  token: string;
  message: string;
}

export type DashboardStatsWeeklyRevenueItem = {
  day: string;
  revenue: number;
};

export interface DashboardStats {
  totalOrdersToday: number;
  revenueToday: number;
  pendingOrders: number;
  completedOrders: number;
  /** @nullable */
  popularItem: string | null;
  totalMenuItems: number;
  weeklyRevenue?: DashboardStatsWeeklyRevenueItem[];
}

export type ListMenuItemsParams = {
category?: string;
available?: boolean;
};

export type ListOrdersParams = {
status?: string;
};