# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shakti Fast Food API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: menu
    description: Menu item operations
  - name: categories
    description: Category operations
  - name: orders
    description: Order operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: available
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"

  /menu/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
    put:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /menu/{id}/availability:
    patch:
      operationId: toggleMenuItemAvailability
      tags: [menu]
      summary: Toggle menu item availability (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "401":
          description: Invalid credentials

  /admin/dashboard:
    get:
      operationId: getDashboardStats
      tags: [admin]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /menu/popular:
    get:
      operationId: getPopularItems
      tags: [menu]
      summary: Get popular/featured menu items
      responses:
        "200":
          description: Popular items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    MenuItem:
      type: object
      required: [id, name, price, category, available, isVeg]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        halfPrice:
          type: ["number", "null"]
        category:
          type: string
        image:
          type: ["string", "null"]
        available:
          type: boolean
        isVeg:
          type: boolean
        isFeatured:
          type: boolean
        orderCount:
          type: integer

    MenuItemInput:
      type: object
      required: [name, price, category, isVeg]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        halfPrice:
          type: ["number", "null"]
        category:
          type: string
        image:
          type: ["string", "null"]
        available:
          type: boolean
        isVeg:
          type: boolean
        isFeatured:
          type: boolean

    MenuItemUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        halfPrice:
          type: ["number", "null"]
        category:
          type: string
        image:
          type: ["string", "null"]
        available:
          type: boolean
        isVeg:
          type: boolean
        isFeatured:
          type: boolean

    AvailabilityUpdate:
      type: object
      required: [available]
      properties:
        available:
          type: boolean

    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: ["string", "null"]
        itemCount:
          type: integer

    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        icon:
          type: ["string", "null"]

    OrderItem:
      type: object
      required: [menuItemId, name, price, quantity]
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        size:
          type: ["string", "null"]

    Order:
      type: object
      required: [id, customerName, phone, items, total, status, orderType, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        phone:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        status:
          type: string
          enum: [received, preparing, ready, completed, cancelled]
        orderType:
          type: string
          enum: [dine-in, takeaway]
        notes:
          type: ["string", "null"]
        tableNumber:
          type: ["string", "null"]
        createdAt:
          type: string

    OrderInput:
      type: object
      required: [customerName, phone, items, orderType]
      properties:
        customerName:
          type: string
        phone:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        orderType:
          type: string
          enum: [dine-in, takeaway]
        notes:
          type: ["string", "null"]
        tableNumber:
          type: ["string", "null"]

    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [received, preparing, ready, completed, cancelled]

    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AuthToken:
      type: object
      required: [token, message]
      properties:
        token:
          type: string
        message:
          type: string

    DashboardStats:
      type: object
      required: [totalOrdersToday, revenueToday, pendingOrders, completedOrders, popularItem, totalMenuItems]
      properties:
        totalOrdersToday:
          type: integer
        revenueToday:
          type: number
        pendingOrders:
          type: integer
        completedOrders:
          type: integer
        popularItem:
          type: ["string", "null"]
        totalMenuItems:
          type: integer
        weeklyRevenue:
          type: array
          items:
            type: object
            required: [day, revenue]
            properties:
              day:
                type: string
              revenue:
                type: number