# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shakti Fast Food API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  name: string;
  slug: string;
  /** @nullable */
  icon?: string | null;
}