# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shakti Fast Food API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsWeeklyRevenueItem } from './dashboardStatsWeeklyRevenueItem';

export interface DashboardStats {
  totalOrdersToday: number;
  revenueToday: number;
  pendingOrders: number;
  completedOrders: number;
  /** @nullable */
  popularItem: string | null;
  totalMenuItems: number;
  weeklyRevenue?: DashboardStatsWeeklyRevenueItem[];
}