# lib/api-zod/src/generated/types/menuItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shakti Fast Food API
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItemInput {
  name: string;
  description?: string;
  price: number;
  /** @nullable */
  halfPrice?: number | null;
  category: string;
  /** @nullable */
  image?: string | null;
  available?: boolean;
  isVeg: boolean;
  isFeatured?: boolean;
}