# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shakti Fast Food API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderOrderType } from './orderOrderType';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  orderType: OrderOrderType;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tableNumber?: string | null;
  createdAt: string;
}