# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shakti Fast Food API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputOrderType } from './orderInputOrderType';
import type { OrderItem } from './orderItem';

export interface OrderInput {
  customerName: string;
  phone: string;
  items: OrderItem[];
  orderType: OrderInputOrderType;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tableNumber?: string | null;
}